Took out test target's test_check dependency...
I have not finished writing that dependency and testing it enough to
have anything in the build file really care about using it.  It is
meant as an optimization for the test target so that in the context of
CI it would not run tests if everything was aready up to date.  I
will test this more thoroughly before I have anyone depend on it.

-Gavin