Now only one place to set versions...
The problem was that I set the version in one place to build, but then
would have to manually modify the post.manif file to post artifacts to
the distribution server.  So now I have it so that there is a new
"post" target that will basically exec the post script in bin, but
before it does this it performs token filtering in the templated
post.manif file (post.manif.tmpl) filling in the module and version
values. This makes it so there is only one place to set the version;
in etc esgf-security.ant.properties.

So now there are three distibution targets.

post - sends specified artifacts to DISTRIBUTION server
post_docs - sends all the generated documentation to the WEB server
publish - sends artifacts (jars) to ARTIFACT repository

This begs the question, "Why even have a manifest file? Just post the
artifacts that are generated in the dist directory, right?" Well, it
may come to that, however, the value of having a manifest is that you
get to explicity state what you want posted.

also I don't feel comfortable with posting anything other than jars to
the artifactory repository, so that has it's own mechanism that is
more tightly integrated with ivy and ant.

The templating and token replacement feature of ant is used for the
post.manif file, the python setup.py file and the python database
initialization script.