Revert "Allowing SQLExceptions to propagate to the clients."
According to me (gavin)
Exceptions should be handled as close to where they can be
semantically handled as possible.  The DAO should deal with exceptions
the best it can before airing it's dirty laundry.  Also the DAO does
not have to necessarily advertise the exception since it is very sql
related and not all DAOs (possible future DAOs) will be backed by a
database.

Indeed truly exceptional conditions will require exceptions thrown but
they should be runtime exceptions and not have to force callers to
deal with.

IMHO... so will put back the exception catching from within the DAO
and if necessary re-throw runtime exceptions encapsulating the SQL
exception.

This reverts commit 607b24dad7df7e3a0eb2def4798f910433af9ec6.