Support for permission approval...
There is an additional field for permissions that indicates if the
permission has been approved.  It will be then up to the administrator
to set the approval flag to grant the user the permisison.

Initially when groups are added they will 'inherit' the value of the
group's automatic_approval flag.  Subsequent to that the admin may
make modifications to that.

(not quite sure what to use visibility for aside from perhaps p2p
visibility in the registry, perhaps)