Explicitly calling xalan's xslt transformer (not leaving up to factory)
This came up as we moved to the Thredds 4.3.16 where they are using
saxon as their xlt transformer and esgf-security is using xalan.
As the factory function is apparently finding saxon first, the code
that we have was throwing an exception and suggesting we use saxon's
implementation.

In order to fix the transformer to the xalan one we now call the
transformer instances specifically.

See:
http://stackoverflow.com/questions/11224850/using-xalan-alongside-saxon

And...
2013-05-27T15:11:38.108 +0200 [     22946][       1] WARN  - esg.orp.app.SAMLAuthorizer - net.sf.saxon.trans.DynamicError: DOMSource cannot be processed: check that saxon8-dom.jar is on the classpath
java.lang.RuntimeException: net.sf.saxon.trans.DynamicError: DOMSource cannot be processed: check that saxon8-dom.jar is on the classpath
        at org.opensaml.xml.util.XMLHelper.prettyPrintXML(XMLHelper.java:882)
        at esg.security.authz.service.impl.SAMLAuthorizationServiceClientSoapImpl.buildAuthorizationRequest(SAMLAuthorizationServiceClientSoapImpl.java:85)
        at esg.orp.app.SAMLAuthorizer.authorize(SAMLAuthorizer.java:56)
        at esg.orp.app.SAMLAuthorizationServiceFilterCollaborator.authorize(SAMLAuthorizationServiceFilterCollaborator.java:55)
        at esg.orp.app.AuthorizationFilter.attemptValidation(AuthorizationFilter.java:68)
        at esg.orp.app.AccessControlFilterTemplate.doFilter(AccessControlFilterTemplate.java:63)
        at org.apache.catalina.core.ApplicationFilterChain.internalDoFilter(ApplicationFilterChain.java:235)

...